+++ tests/testthat/test-grid_hexametrics.R
@@ -2,12 +2,12 @@ context("grid_hexametrics")
 
 las = lidR:::dummy_las(2000)
 
-test_that("grid_metrics works", {
+test_that("grid_hexametrics works", {
   x = grid_hexametrics(las, mean(Z))
   expect_equal(dim(x), c(39,3))
 })
 
-test_that("grid_metrics space hexa properly", {
+test_that("grid_hexametrics space hexa properly", {
   x = grid_hexametrics(las, mean(Z))
   y = unique(diff(sort(unique(x$X))))
   expect_equal(length(y), 1)
@@ -16,17 +16,17 @@ test_that("grid_metrics space hexa properly", {
   expect_true(100 %in% x$X)
 })
 
-test_that("grid_metrics debug mode works", {
-  expect_error(grid_hexametrics(las, LAD(Z)))
+test_that("grid_hexametrics debug mode works", {
+  expect_error(grid_hexametrics(las, LAD(Z), debug = TRUE))
 })
 
-test_that("grid_metrics return an error if splitline and no flightlineID", {
+test_that("grid_hexametrics return an error if splitline and no flightlineID", {
   expect_error(grid_hexametrics(las, mean(Z), splitlines = T))
 })
 
 las@data[, flightlineID := c(rep(1,500), rep(2,500))]
 
-test_that("grid_metrics splitline work", {
+test_that("grid_hexametrics splitline work", {
   x = grid_hexametrics(las, mean(Z), splitlines = T)
   expect_equal(dim(x), c(78,4))
 })