+++ tests/testthat/test-grid_metrics.R
@@ -8,7 +8,7 @@ test_that("grid_metrics works", {
 })
 
 test_that("grid_metrics debug mode works", {
-  expect_error(grid_metrics(las, LAD(Z)))
+  expect_error(grid_metrics(las, LAD(Z), debug = TRUE))
 })
 
 test_that("grid_metrics return an error if splitline and no flightlineID", {